requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.1
attrs==21.4.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==4.1.0
blinker==1.5
Brotli==1.0.9
brotlipy==0.7.0
cachetools==4.2.2
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
contourpy==1.0.5
cryptography==38.0.3
cycler==0.11.0
d2l==0.17.5
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
fonttools==4.38.0
frozenlist==1.2.0
google-auth==2.6.0
google-auth-oauthlib==0.4.4
grpcio==1.42.0
idna==2.10
importlib-metadata==4.11.3
importlib-resources==5.2.0
ipykernel==6.9.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.3
kaitaistruct==0.10
kiwisolver==1.4.4
kombu==5.2.4
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.1.0
nest-asyncio==1.5.5
netaddr==0.8.0
notebook==6.4.12
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==22.0
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pip==22.1.2
prometheus-client==0.14.1
prompt-toolkit==3.0.20
protobuf==3.20.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydivert==2.1.0
Pygments==2.11.2
PyJWT==2.4.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.2.0
qtconsole==5.3.2
QtPy==2.2.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
selenium-wire==5.1.0
Send2Trash==1.8.0
setuptools==62.3.2
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
terminado==0.13.1
testpath==0.6.0
toml==0.10.2
torch==1.12.1
torch-tb-profiler==0.4.0
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.3.0
urllib3==1.26.11
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==4.0.3
win-inet-pton==1.1.0
wincertstore==0.2
yarl==1.8.1
zipp==3.8.0
zstandard==0.19.0
